Allow to use UUID value (not pointer)
Because basic accessors are defined against pointers, they can't be
called on simple value, which forces to call (&myid).Version().

This commit makes these method accept both value and pointers.

Note: there will be unnecessary copy when calling these methods,
but the size of an UUID is really small and these performance
considerations should be let to the compiler.